back/src/auth-user/requirements.txt
psycopg2-binary
pyotp
cryptography
python-dotenv
